tests/wpt/tests/fullscreen/rendering/fullscreen-pseudo-class-support.html

<!DOCTYPE html>
<title>:fullscreen pseudo-class support</title>
<link rel="help" href="https://fullscreen.spec.whatwg.org/#:fullscreen-pseudo-class">
<link rel="help" href="https://dom.spec.whatwg.org/#dom-element-matches">
<link rel="help" href="https://dom.spec.whatwg.org/#dom-parentnode-queryselector">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
// This is a separate test so that support for :fullscreen can be detected
// without actually going fullscreen. If it were not supported, then this test
// would fail.
test(t => {
  // precondition is to throw for unknown pseudo-classes:
  assert_throws_dom("SyntaxError", () => document.body.matches(':halfscreen'));
  assert_throws_dom("SyntaxError", () => document.querySelector(':halfscreen'));
  // the actual test:
  assert_false(document.body.matches(':fullscreen'));
  assert_equals(document.querySelector(':fullscreen'), null);
});
</script>
